Crate hyphenation

Reexports

pub use hyphenator::{Hyphenation, Standard};

pub use language::{Language, Corpus};

pub use load::{set_pattern_folder, language as load};

Modules

exception
Data structures and methods for parsing and applying exceptions, which assign predetermined scores to specific words.

hyphenator
Hyphenating iterators.

language
Languages we can hyphenate and their default parameters, as provided by the TeX hyph-utf8 package.

load
IO operations for pattern and exception data provided by hyph-UTF8 and stored in the patterns folder.

pattern
Data structures and methods for parsing and applying Knuth-Liang hyphenation patterns.